+++ example/si_constants.cpp
@@ -20,30 +20,21 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#pragma once
+#include <units/format.h>
+#include <units/isq/si/constants.h>
+#include <iostream>
 
-#include <units/derived_dimension.h>
-#include <units/magnitude.h>
-
-namespace units::detail {
-
-// compatible_units
-template<typename ExpList, Unit... Us>
-inline constexpr bool compatible_units = false;
-
-template<typename... Es, Unit... Us>
-inline constexpr bool compatible_units<exponent_list<Es...>, Us...> = (UnitOf<Us, typename Es::dimension> && ...);
-
-// derived_scaled_unit
-
-template<Unit... Us, typename... Es>
-constexpr Magnitude auto derived_mag(exponent_list<Es...>)
+int main()
 {
-  return (as_magnitude<1>() * ... *
-          pow<ratio{Es::num, Es::den}>(Us::mag / dimension_unit<typename Es::dimension>::mag));
+  using namespace units::isq::si::si2019;
+
+  std::cout << "The seven defining constants of the SI and the seven corresponding units they define:\n";
+  std::cout << STD_FMT::format("- hyperfine transition frequency of Cs: {:%.0Q %q}\n",
+                               hyperfine_structure_transition_frequency<>);
+  std::cout << STD_FMT::format("- speed of light in vacuum:             {:%.0Q %q}\n", speed_of_light<>);
+  std::cout << STD_FMT::format("- Planck constant:                      {}\n", planck_constant<>);
+  std::cout << STD_FMT::format("- elementary charge:                    {}\n", elementary_charge<>);
+  std::cout << STD_FMT::format("- Boltzmann constant:                   {}\n", boltzmann_constant<>);
+  std::cout << STD_FMT::format("- Avogadro constant:                    {}\n", avogadro_constant<>);
+  std::cout << STD_FMT::format("- luminous efficacy:                    {}\n", luminous_efficacy<>);
 }
-
-template<DerivedDimension D, Unit... Us>
-using derived_scaled_unit = scaled_unit<derived_mag<Us...>(typename D::recipe()), typename D::coherent_unit::reference>;
-
-}  // namespace units::detail