+++ test/unit_test/static/quantity_test.cpp
@@ -288,8 +288,10 @@ static_assert(get_length_derived_quantity() == 1_q_m);
 /////////
 
 #if UNITS_COMP_GCC >= 11 || UNITS_COMP_CLANG > 15
-static_assert(std::is_same_v<decltype(units::aliases::isq::si::m(123))::rep, int>);
-static_assert(std::is_same_v<decltype(units::aliases::isq::si::m(123.))::rep, double>);
+static_assert(std::is_same_v<decltype(aliases::isq::si::m(123))::rep, int>);
+static_assert(std::is_same_v<decltype(aliases::isq::si::m(123.))::rep, double>);
+static_assert(
+  std::is_same_v<decltype(aliases::isq::si::m3(aliases::isq::si::m2{2} * aliases::isq::si::m{3}))::rep, int>);
 #endif
 
 static_assert(is_same_v<decltype(quantity{length<metre, int>(123)}), length<metre, int>>);